Move the file info boxes section to the end of the page
The keyword index is probably the first thing they want to see, and
this also has the advantage that jumping from the TOC to an index will
always jump to the right location, without having to do the hack of
adding 40 <br> or so at the end of the page. (Well, now the problem is
just with the "File info by folder" section, though.)

Also remove the upwards pointing arrow symbol as it's now wrong in
addition to being redundant, the "i" symbol is enough.