Add `make-xmlhub-indexer-release` binary, replace git-version with build.rs
Add a `make-xmlhub-indexer-release` binary that carries out all the
details for a release, including publishing of the binary.

This replaces the git-version crate, because cargo wouldn't recognize
when a git tag is created and hence the published binary would refer
to a version before the release. There would be a possible hack to
watch the directory containing tag files (when not packed), but that
may not be reliable and would require patching the git-version crate.
The only way to force a rebuild is via `cargo clean`, and that would
be overkill. Thus instead use a `build.rs` to let cargo watch out for
a file `.released_version`, and have the release process update that
to trigger a rebuild when it creates a tag. It also does use the
mentioned hack which does make rebuilds work for manual Git
manipulations, when using the git binary as of today.