read_xml: re-bundle backing and Document
This allows the user to again move the document backing along when
moving the parsed `Document` (i.e. stops the user having to deal with
the self-referential structure issue) by having both bundled in the
returned `XMLDocument`.

The "trick" that needed solving here is to not try to implement the
`Deref` trait but have our own `document` accessor instead
(`borrow_document` is not `pub`): `Deref` would not allow to pass the
required lifetime parameter to `Document`.