xmlhub: move the state folder out of .git
It and .git directly both having "logs" directory, as well as the fact
that future use with multiple repositories would not have a `.git`
folder at the toplevel, makes it clear that this was not only a hack
but a bad one.

Just have to add this folder to `.gitignore` now.