xmlhub: from info box, link back to index via separate links
Solve the problem with value having links themselves, leading to
layered links of which one is not clickable (nor visible).

(Makes extensive use of `Flat`, to do away with `<div>` elements
breaking up lines (and `<span>` is not allowed there). But
additionally hack cutting of surplus `<br>` at the end of the SoftPre
output. Bah, TODO.)